Extract to src/components/viewerlib/
