+++ src/components/viewer.ts
@@ -7,99 +7,13 @@ import type {Extension} from '../main'
 import type {SyncTeXRecordForward} from './locator'
 import {openWebviewPanel} from '../utils/webview'
 
-import type {ClientRequest, ServerResponse, PanelRequest, PdfViewerState} from '../../viewer/components/protocol'
-import {escapeHtml} from '../utils/utils'
+import type {ClientRequest, ServerResponse, PdfViewerState} from '../../viewer/components/protocol'
 import {getCurrentThemeLightness} from '../utils/theme'
 
+import {Client} from './viewerlib/client'
+import {PdfViewerPanel, PdfViewerPanelSerializer} from './viewerlib/pdfviewerpanel'
 export {PdfViewerHookProvider} from './viewerlib/pdfviewerhook'
 
-class Client {
-    readonly viewer: 'browser' | 'tab'
-    readonly websocket: ws
-    private readonly disposables = new Set<vscode.Disposable>()
-
-    constructor(viewer: 'browser' | 'tab', websocket: ws) {
-        this.viewer = viewer
-        this.websocket = websocket
-        this.websocket.on('close', () => {
-            this.disposeDisposables()
-        })
-    }
-
-    private disposeDisposables() {
-        vscode.Disposable.from(...this.disposables).dispose()
-        this.disposables.clear()
-    }
-
-    onDidDispose(cb: () => unknown) {
-        this.disposables.add( { dispose: cb } )
-    }
-
-    send(message: ServerResponse) {
-        this.websocket.send(JSON.stringify(message))
-    }
-}
-
-class PdfViewerPanel {
-    readonly webviewPanel: vscode.WebviewPanel
-    readonly pdfFileUri: vscode.Uri
-    private _state?: PdfViewerState
-
-    constructor(pdfFileUri: vscode.Uri, panel: vscode.WebviewPanel) {
-        this.pdfFileUri = pdfFileUri
-        this.webviewPanel = panel
-        panel.webview.onDidReceiveMessage((msg: PanelRequest) => {
-            switch(msg.type) {
-                case 'state': {
-                    this._state = msg.state
-                    break
-                }
-                default: {
-                    break
-                }
-            }
-        })
-    }
-
-    get state() {
-        return this._state
-    }
-
-}
-
-class PdfViewerPanelSerializer implements vscode.WebviewPanelSerializer {
-    private readonly extension: Extension
-
-    constructor(extension: Extension) {
-        this.extension = extension
-    }
-
-    async deserializeWebviewPanel(panel: vscode.WebviewPanel, argState: {state: PdfViewerState}): Promise<void> {
-        this.extension.logger.addLogMessage(`Restoring the PDF viewer at the column ${panel.viewColumn} from the state: ${JSON.stringify(argState)}`)
-        const state = argState.state
-        let pdfFileUri: vscode.Uri | undefined
-        if (state.path) {
-            pdfFileUri = vscode.Uri.file(state.path)
-        } else if (state.pdfFileUri) {
-            pdfFileUri = vscode.Uri.parse(state.pdfFileUri, true)
-        }
-        if (!pdfFileUri) {
-            this.extension.logger.addLogMessage('Error of restoring PDF viewer: the path of PDF file is undefined.')
-            panel.webview.html = '<!DOCTYPE html> <html lang="en"><meta charset="utf-8"/><br>The path of PDF file is undefined.</html>'
-            return
-        }
-        if (! await this.extension.lwfs.exists(pdfFileUri)) {
-            const s = escapeHtml(pdfFileUri.toString())
-            this.extension.logger.addLogMessage(`Error of restoring PDF viewer: file not found ${pdfFileUri.toString(true)}.`)
-            panel.webview.html = `<!DOCTYPE html> <html lang="en"><meta charset="utf-8"/><br>File not found: ${s}</html>`
-            return
-        }
-        panel.webview.html = await this.extension.viewer.getPDFViewerContent(pdfFileUri)
-        this.extension.viewer.initiatePdfViewerPanel(pdfFileUri, panel)
-        return
-    }
-
-}
 
 export class Viewer {
     private readonly extension: Extension
